+++ packages/jupyterlab-drawio/src/io.ts
@@ -16,7 +16,7 @@ import { LabIcon } from '@jupyterlab/ui-components';
 import DRAWIO_ICON_SVG from '../style/img/drawio.svg';
 
 import {
-  IDiagramManager,
+  IFormat,
   DRAWIO_ICON_CLASS_RE,
   TEXT_FACTORY,
   BINARY_FACTORY,
@@ -39,13 +39,13 @@ export const drawioPngIcon = new LabIcon({
   svgstr: DRAWIO_ICON_SVG.replace(DRAWIO_ICON_CLASS_RE, 'jp-icon-contrast0'),
 });
 
-export const XML_NATIVE: IDiagramManager.IFormat = {
+export const XML_NATIVE: IFormat = {
   ext: '.dio',
   format: 'text',
   icon: drawioIcon,
   key: 'drawio',
   label: 'Diagram',
-  mimetype: 'application/dio',
+  mimetype: 'application/x-drawio',
   name: 'dio',
   modelName: 'text',
   factoryName: TEXT_FACTORY,
@@ -54,15 +54,19 @@ export const XML_NATIVE: IDiagramManager.IFormat = {
   isEditable: true,
   isText: true,
   isDefault: true,
+  isTransformed: false,
+  exporter: async (drawio, key, settings) => {
+    return await drawio.adapter.toXML();
+  },
 };
 
-export const XML_LEGACY: IDiagramManager.IFormat = {
+export const XML_LEGACY: IFormat = {
   ext: '.drawio',
   format: 'text',
   icon: drawioIcon,
   key: 'dio',
   label: 'Diagram (mxgraph)',
-  mimetype: 'application/mxgraph',
+  mimetype: 'application/x-mxgraph',
   name: 'dio-legacy',
   modelName: 'text',
   factoryName: `${TEXT_FACTORY} (Legacy)`,
@@ -70,9 +74,10 @@ export const XML_LEGACY: IDiagramManager.IFormat = {
   isEditable: true,
   isText: true,
   isDefault: true,
+  isTransformed: false,
 };
 
-export const SVG_PLAIN: IDiagramManager.IFormat = {
+export const SVG_PLAIN: IFormat = {
   ext: '.svg',
   format: 'text',
   factoryName: `${TEXT_FACTORY} (SVG)`,
@@ -85,9 +90,10 @@ export const SVG_PLAIN: IDiagramManager.IFormat = {
   save: unbase64SVG,
   isExport: true,
   isText: true,
+  isTransformed: true,
 };
 
-export const SVG_EDITABLE: IDiagramManager.IFormat = {
+export const SVG_EDITABLE: IFormat = {
   ...SVG_PLAIN,
   factoryName: `${TEXT_FACTORY} (Editable SVG)`,
   ext: '.dio.svg',
@@ -97,9 +103,10 @@ export const SVG_EDITABLE: IDiagramManager.IFormat = {
   pattern: '^.*.dio.svg$',
   isEditable: true,
   isDefault: true,
+  isTransformed: true,
 };
 
-export const PNG_PLAIN: IDiagramManager.IFormat = {
+export const PNG_PLAIN: IFormat = {
   ext: '.png',
   format: 'base64',
   modelName: 'base64',
@@ -112,9 +119,10 @@ export const PNG_PLAIN: IDiagramManager.IFormat = {
   save: stripDataURI,
   isBinary: true,
   isExport: true,
+  isTransformed: true,
 };
 
-export const PNG_EDITABLE: IDiagramManager.IFormat = {
+export const PNG_EDITABLE: IFormat = {
   ...PNG_PLAIN,
   factoryName: `${BINARY_FACTORY} (Editable PNG)`,
   ext: '.dio.png',