+++ packages/jupyterlab-drawio/src/plugin.ts
@@ -36,6 +36,7 @@ import { NS, PLUGIN_ID } from '.';
 
 import { IDiagramManager, CommandIds, DEBUG } from './tokens';
 import { DiagramManager } from './manager';
+import { RenderedDiagram } from './mime';
 
 /**
  * The editor tracker extension.
@@ -55,7 +56,7 @@ const plugin: JupyterFrontEndPlugin<IDiagramManager> = {
   autoStart: true,
 };
 
-export default plugin;
+export default [plugin];
 
 function activate(
   app: JupyterLab,
@@ -112,6 +113,6 @@ function activate(
     menu.fileMenu.newMenu.addGroup([{ command: CommandIds.createNew }], 40);
   }
 
-  // this is very odd, and probably can't be reused. Use the manager pattern?
+  RenderedDiagram.manager = manager;
   return manager;
 }