+++ py_packages/jupyter-drawio-export/src/jupyter_drawio_export/handlers.py
@@ -50,8 +50,7 @@ class ProvisionHandler(BaseHandler):
 
 
 def add_handlers(nbapp):
-    """ Add drawio routes to the notebook server web application
-    """
+    """Add drawio routes to the notebook server web application"""
     ns_url = ujoin(nbapp.base_url, "drawio")
     pdf_url = ujoin(ns_url, "export", r"(?P<url>.*)")
     status_url = ujoin(ns_url, "status")