+++ py_packages/jupyter-drawio-export/src/jupyter_drawio_export/serverextension.py
@@ -23,8 +23,7 @@ from .manager import DrawioExportManager
 
 
 def load_jupyter_server_extension(nbapp):
-    """ create a DrawioExportManager and add handlers
-    """
+    """create a DrawioExportManager and add handlers"""
     nbapp.add_traits(drawio_manager=traitlets.Instance(DrawioExportManager))
     manager = nbapp.drawio_manager = DrawioExportManager(parent=nbapp, log=nbapp.log)
     manager.initialize()