+++ scripts/atest.py
@@ -1,126 +1,113 @@
-""" Run acceptance tests with robot framework
-"""
-# pylint: disable=broad-except
 import os
 import shutil
+import subprocess
 import sys
 import time
-from os.path import join
-
-import robot
 
 from . import project as P
 
-# sets of tags to be combined with AND, marked as non-critical
-NON_CRITICAL = []
+PABOT_DEFAULTS = [
+    "--testlevelsplit",
+    "--processes",
+    "4",
+    "--artifactsinsubfolders",
+    "--artifacts",
+    "png,log,txt,drawio,xml",
+]
 
 
-def get_stem(attempt, extra_args):
-    stem = "_".join([P.PLATFORM, P.PY_MAJOR, str(attempt)]).replace(".", "_").lower()
+def run_tests(attempt=0, extra_args=None):
+    extra_args = extra_args or []
 
-    if "--dryrun" in extra_args:
-        stem = f"dry_run_{stem}"
+    stem = P.get_atest_stem(attempt=attempt, extra_args=extra_args)
+    out_dir = P.ATEST_OUT / stem
 
-    return stem
-
-
-def atest(attempt, extra_args):
-    """ perform a single attempt of the acceptance tests
-    """
-
-    if "FIREFOX_BINARY" not in os.environ:
-        os.environ["FIREFOX_BINARY"] = shutil.which("firefox")
-
-        prefix = os.environ.get("CONDA_PREFIX")
-
-        if prefix:
-            app_dir = join(prefix, "bin", "FirefoxApp")
-            os.environ["FIREFOX_BINARY"] = {
-                "Windows": join(prefix, "Library", "bin", "firefox.exe"),
-                "Linux": join(app_dir, "firefox"),
-                "Darwin": join(app_dir, "Contents", "MacOS", "firefox"),
-            }[P.PLATFORM]
-
-    print("Will use firefox at", os.environ["FIREFOX_BINARY"])
-
-    assert os.path.exists(os.environ["FIREFOX_BINARY"])
-
-    stem = get_stem(attempt, extra_args)
-
-    for non_critical in NON_CRITICAL:
-        extra_args += ["--noncritical", "AND".join(non_critical)]
-
-    if attempt != 1:
-        previous = P.ATEST_OUT / f"{get_stem(attempt - 1, extra_args)}.robot.xml"
+    if attempt > 1:
+        prev_stem = P.get_atest_stem(attempt=attempt - 1, extra_args=extra_args)
+        previous = P.ATEST_OUT / prev_stem / P.ATEST_OUT_XML
         if previous.exists():
             extra_args += ["--rerunfailed", str(previous)]
 
-    out_dir = P.ATEST_OUT / stem
+    runner = ["pabot", *PABOT_DEFAULTS]
+
+    if "--dryrun" in extra_args:
+        runner = ["robot"]
 
     args = [
+        *runner,
+        *extra_args,
         "--name",
-        f"{P.PLATFORM}{P.PY_MAJOR}",
+        f"""{P.PLATFORM} py{P.PY_MAJOR}s""",
         "--outputdir",
         out_dir,
-        "--output",
-        P.ATEST_OUT / f"{stem}.robot.xml",
-        "--log",
-        P.ATEST_OUT / f"{stem}.log.html",
-        "--report",
-        P.ATEST_OUT / f"{stem}.report.html",
-        "--xunitskipnoncritical",
-        "--xunit",
-        P.ATEST_OUT / f"{stem}.xunit.xml",
         "--variable",
         f"OS:{P.PLATFORM}",
         "--variable",
         f"PY:{P.PY_MAJOR}",
         "--randomize",
         "all",
-        *(extra_args or []),
-        P.ATEST,
+        "--xunitskipnoncritical",
+        "--xunit",
+        ".".join(["xunit", "xml"]),
+        ".",
     ]
 
-    print("Robot Arguments\n", " ".join(["robot"] + list(map(str, args))))
-
-    os.chdir(P.ATEST)
-
     if out_dir.exists():
-        print("trying to clean out {}".format(out_dir))
+        print(">>> trying to clean out {}".format(out_dir), flush=True)
         try:
             shutil.rmtree(out_dir)
         except Exception as err:
-            print("Error deleting {}, hopefully harmless: {}".format(out_dir, err))
+            print(
+                "... error, hopefully harmless: {}".format(err),
+                flush=True,
+            )
+
+    if not out_dir.exists():
+        print(">>> trying to prepare output directory: {}".format(out_dir), flush=True)
+        try:
+            out_dir.mkdir(parents=True)
+        except Exception as err:
+            print(
+                "... Error, hopefully harmless: {}".format(err),
+                flush=True,
+            )
+
+    str_args = [*map(str, args)]
+    print(">>> ", " ".join(str_args), flush=True)
+
+    proc = subprocess.Popen(str_args, cwd=P.ATEST)
 
     try:
-        robot.run_cli(list(map(str, args)))
-        return 0
-    except SystemExit as err:
-        return err.code
-    finally:
-        for dot_dir in out_dir.rglob(".*/"):
-            if dot_dir.is_dir():
-                print("cleaning", dot_dir, flush=True)
-                shutil.rmtree(dot_dir)
+        return proc.wait()
+    except KeyboardInterrupt:
+        proc.kill()
+        return proc.wait()
 
 
-def attempt_atest_with_retries(*extra_args):
-    """ retry the robot tests a number of times
-    """
+def attempt_atest_with_retries(extra_args=None):
+    """retry the robot tests a number of times"""
+    extra_args = list(extra_args or [])
     attempt = 0
     error_count = -1
 
     retries = int(os.environ.get("ATEST_RETRIES") or "0")
+    extra_args += os.environ.get("ATEST_ARGS", "").split()
 
     while error_count != 0 and attempt <= retries:
         attempt += 1
-        print("attempt {} of {}...".format(attempt, retries + 1))
+        print("attempt {} of {}...".format(attempt, retries + 1), flush=True)
         start_time = time.time()
-        error_count = atest(attempt=attempt, extra_args=list(extra_args))
-        print(error_count, "errors in", int(time.time() - start_time), "seconds")
+        error_count = run_tests(attempt=attempt, extra_args=extra_args)
+        print(
+            error_count,
+            "errors in",
+            int(time.time() - start_time),
+            "seconds",
+            flush=True,
+        )
 
     return error_count
 
 
 if __name__ == "__main__":
-    sys.exit(attempt_atest_with_retries(*sys.argv[1:]))
+    sys.exit(attempt_atest_with_retries(extra_args=sys.argv[1:]))